html/browsers/the-window-object/apis-for-creating-and-navigating-browsing-contexts-by-name/open-features-optional-004.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML: window.open `features`: tokenization -- legacy size features `innerheight`, `innerwidth`</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/#apis-for-creating-and-navigating-browsing-contexts-by-name">

<!-- user agents are not required to support open features other than `noopener`
     and on some platforms position and size features don't make sense -->
<meta name="flags" content="may" />

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/PrefixedPostMessage.js"></script>
<script>
var windowURL = 'resources/message-opener.html';

[ 'innerwidth=142',
  ' innerwidth = 142',
  'innerwidth==142',
  '\ninnerwidth= 142',
  ',innerwidth=142,,',
  'INNERWIDTH=142',
  'innerWidth=142'
].forEach((features, idx, arr) => {
  async_test(t => {
    var prefixedMessage = new PrefixedMessageTest();
    prefixedMessage.onMessage(t.step_func_done((data, e) => {
      e.source.close();
      assert_equals(data.width, 142);
    }));
    var win = window.open(prefixedMessage.url(windowURL), '', features);
  }, `"${features}" (${idx + 1} of ${arr.length}) should set width of opened window`);
});

[ 'innerheight=142',
  ' innerheight = 142',
  'innerheight==142',
  '\ninnerheight= 142',
  ',innerheight=142,,',
  'INNERHEIGHT=142',
  'innerHeight=142'
].forEach((features, idx, arr) => {
  async_test(t => {
    var prefixedMessage = new PrefixedMessageTest();
    prefixedMessage.onMessage(t.step_func_done((data, e) => {
      e.source.close();
      assert_equals(data.height, 142);
    }));
    var win = window.open(prefixedMessage.url(windowURL), '', features);
  }, `"${features}" (${idx + 1} of ${arr.length}) should set height of opened window`);
});

</script>
